Continue

+++ models/products.model.js
@@ -1,29 +1,44 @@
 // Import Product Model Object
 
-const { productModel, categoryModel, mongoose } = require("../models/all.models");
+const { productModel, categoryModel, adminModel, mongoose } = require("../models/all.models");
 
-async function addNewProduct(productInfo) {
+async function addNewProduct(authorizationId, productInfo) {
     try {
-        const product = await productModel.findOne({ name: productInfo.name, category: productInfo.category });
-        if (!product) {
-            const category = await categoryModel.findOne({ name: productInfo.category });
-            if (category) {
-                const newProductInfo = new productModel(productInfo);
-                await newProductInfo.save();
+        const admin = await adminModel.findById(authorizationId);
+        if (admin){
+            if ((new mongoose.Types.ObjectId(authorizationId)).equals(admin._id) && !admin.isBlocked) {
+                const product = await productModel.findOne({ name: productInfo.name, category: productInfo.category });
+                if (!product) {
+                    const category = await categoryModel.findOne({ name: productInfo.category });
+                    if (category) {
+                        const newProductInfo = new productModel(productInfo);
+                        await newProductInfo.save();
+                        return {
+                            msg: "Adding New Product Process Has Been Successfuly !!",
+                            error: false,
+                            data: {},
+                        }
+                    }
+                    return {
+                        msg: "Sorry, This Category Is Not Exist !!",
+                        error: true,
+                        data: {},
+                    }
+                }
                 return {
-                    msg: "Adding New Product Process Has Been Successfuly !!",
-                    error: false,
+                    msg: "Sorry, This Product Is Already Exist !!",
+                    error: true,
                     data: {},
                 }
             }
             return {
-                msg: "Sorry, This Category Is Not Exist !!",
+                msg: "Sorry, Permission Denied !!",
                 error: true,
                 data: {},
             }
         }
         return {
-            msg: "Sorry, This Product Is Already Exist !!",
+            msg: "Sorry, This Admin Is Not Exist !!",
             error: true,
             data: {},
         }
@@ -33,25 +48,40 @@ async function addNewProduct(productInfo) {
     }
 }
 
-async function addingNewImagesToProductGallery(productId, newGalleryImagePaths) {
+async function addingNewImagesToProductGallery(authorizationId, productId, newGalleryImagePaths) {
     try{
-        const product = await productModel.findById(productId);
-        if (product) {
-            await productModel.updateOne({ _id: productId },
-            {
-                galleryImagesPaths: product.galleryImagesPaths.concat(newGalleryImagePaths),
-            });
+        const admin = await adminModel.findById(authorizationId);
+        if (admin){
+            if ((new mongoose.Types.ObjectId(authorizationId)).equals(admin._id) && !admin.isBlocked) {
+                const product = await productModel.findById(productId);
+                if (product) {
+                    await productModel.updateOne({ _id: productId },
+                    {
+                        galleryImagesPaths: product.galleryImagesPaths.concat(newGalleryImagePaths),
+                    });
+                    return {
+                        msg: "Adding New Images To Product Gallery Process Has Been Successfuly !!",
+                        error: false,
+                        data: {
+                            newGalleryImagePaths,
+                        },
+                    }
+                }
+                return {
+                    msg: "Sorry, This Product Is Not Found !!",
+                    error: false,
+                    data: {},
+                }
+            }
             return {
-                msg: "Adding New Images To Product Gallery Process Has Been Successfuly !!",
-                error: false,
-                data: {
-                    newGalleryImagePaths,
-                },
+                msg: "Sorry, Permission Denied !!",
+                error: true,
+                data: {},
             }
         }
         return {
-            msg: "Sorry, This Product Is Not Found !!",
-            error: false,
+            msg: "Sorry, This Admin Is Not Exist !!",
+            error: true,
             data: {},
         }
     }
@@ -131,23 +161,38 @@ async function getRelatedProductsInTheProduct(productId) {
     }
 }
 
-async function deleteProduct(productId) {
+async function deleteProduct(authorizationId, productId) {
     try {
-        const productInfo = await productModel.findOneAndDelete({
-            _id: productId,
-        });
-        if (productInfo) {
+        const admin = await adminModel.findById(authorizationId);
+        if (admin){
+            if ((new mongoose.Types.ObjectId(authorizationId)).equals(admin._id) && !admin.isBlocked) {
+                const productInfo = await productModel.findOneAndDelete({
+                    _id: productId,
+                });
+                if (productInfo) {
+                    return {
+                        msg: "Deleting Product Process Has Been Successfuly !!",
+                        error: false,
+                        data: {
+                            deletedProductPath: productInfo.imagePath,
+                            galleryImagePathsForDeletedProduct: productInfo.galleryImagesPaths,
+                        },
+                    }
+                }
+                return {
+                    msg: "Sorry, This Product Is Not Exist !!",
+                    error: true,
+                    data: {},
+                }
+            }
             return {
-                msg: "Deleting Product Process Has Been Successfuly !!",
-                error: false,
-                data: {
-                    deletedProductPath: productInfo.imagePath,
-                    galleryImagePathsForDeletedProduct: productInfo.galleryImagesPaths,
-                },
+                msg: "Sorry, Permission Denied !!",
+                error: true,
+                data: {},
             }
         }
         return {
-            msg: "Sorry, This Product Is Not Exist !!",
+            msg: "Sorry, This Admin Is Not Exist !!",
             error: true,
             data: {},
         }
@@ -159,73 +204,34 @@ async function deleteProduct(productId) {
 
 async function deleteImageFromProductGallery(productId, galleryImagePath) {
     try{
-        const product = await productModel.findById(productId);
-        if (product) {
-            await productModel.updateOne({ _id: productId }, {
-                galleryImagesPaths: product.galleryImagesPaths.filter((path) => galleryImagePath !== path)
-            });
-            return {
-                msg: "Deleting Image From Product Gallery Process Has Been Successfully !!",
-                error: false,
-                data: {},
-            }
-        }
-        return {
-            msg: "Sorry, This Product Is Not Exist !!",
-            error: true,
-            data: {},
-        }
-    }
-    catch(err) {
-        throw Error(err);
-    }
-}
-
-async function updateProduct(productId, newData) {
-    try {
-        const product = await productModel.findOneAndUpdate({ _id: productId }, { ...newData });
-        if (product) {
-            return {
-                msg: "Updating Product Process Has Been Successful !!",
-                error: false,
-                data: {},
-            }
-        }
-        return {
-            msg: "Sorry, This Product Is Not Exist !!",
-            error: true,
-            data: {},
-        }
-    }
-    catch (err) {
-        throw Error(err);
-    }
-}
-
-async function updateProductGalleryImage(productId, oldGalleryImagePath, newGalleryImagePath) {
-    try{
-        const product = await productModel.findById(productId);
-        if (product) {
-            const galleryImagePathIndex = product.galleryImagesPaths.findIndex((galleryImagePath) => galleryImagePath === oldGalleryImagePath);
-            if (galleryImagePathIndex >= 0) {
-                product.galleryImagesPaths[galleryImagePathIndex] = newGalleryImagePath;
-                await productModel.updateOne({ _id: productId }, {
-                    galleryImagesPaths: product.galleryImagesPaths
-                });
+        const admin = await adminModel.findById(authorizationId);
+        if (admin){
+            if ((new mongoose.Types.ObjectId(authorizationId)).equals(admin._id) && !admin.isBlocked) {
+                const product = await productModel.findById(productId);
+                if (product) {
+                    await productModel.updateOne({ _id: productId }, {
+                        galleryImagesPaths: product.galleryImagesPaths.filter((path) => galleryImagePath !== path)
+                    });
+                    return {
+                        msg: "Deleting Image From Product Gallery Process Has Been Successfully !!",
+                        error: false,
+                        data: {},
+                    }
+                }
                 return {
-                    msg: "Updating Product Galley Image Process Has Been Successfully !!",
-                    error: false,
+                    msg: "Sorry, This Product Is Not Exist !!",
+                    error: true,
                     data: {},
                 }
             }
             return {
-                msg: "Sorry, This Path Is Not Found !!",
+                msg: "Sorry, Permission Denied !!",
                 error: true,
                 data: {},
             }
         }
         return {
-            msg: "Sorry, This Product Is Not Exist !!",
+            msg: "Sorry, This Admin Is Not Exist !!",
             error: true,
             data: {},
         }
@@ -235,25 +241,125 @@ async function updateProductGalleryImage(productId, oldGalleryImagePath, newGall
     }
 }
 
-async function updateProductImage(productId, newProductImagePath) {
-    try{
-        const product = await productModel.findById(productId);
-        if (product) {
-            await productModel.updateOne({ _id: productId }, {
-                imagePath: newProductImagePath,
-            });
+async function updateProduct(authorizationId, productId, newData) {
+    try {
+        const admin = await adminModel.findById(authorizationId);
+        if (admin){
+            if ((new mongoose.Types.ObjectId(authorizationId)).equals(admin._id) && !admin.isBlocked) {
+                const product = await productModel.findOneAndUpdate({ _id: productId }, { ...newData });
+                if (product) {
+                    return {
+                        msg: "Updating Product Process Has Been Successfully !!",
+                        error: false,
+                        data: {},
+                    }
+                }
+                return {
+                    msg: "Sorry, This Product Is Not Exist !!",
+                    error: true,
+                    data: {},
+                }
+            }
             return {
-                msg: "Change Product Image Process Has Been Successfully !!",
-                error: false,
-                data: {
-                    deletedProductImagePath: product.imagePath,
-                },
+                msg: "Sorry, Permission Denied !!",
+                error: true,
+                data: {},
             }
         }
         return {
-            msg: "Sorry, This Product Is Not Exist !!",
+            msg: "Sorry, This Admin Is Not Exist !!",
             error: true,
-            data: product.imagePath,
+            data: {},
+        }
+    }
+    catch (err) {
+        console.log(err);
+        throw Error(err);
+    }
+}
+
+async function updateProductGalleryImage(authorizationId, productId, oldGalleryImagePath, newGalleryImagePath) {
+    try{
+        const admin = await adminModel.findById(authorizationId);
+        if (admin){
+            if ((new mongoose.Types.ObjectId(authorizationId)).equals(admin._id) && !admin.isBlocked) {
+                const product = await productModel.findById(productId);
+                if (product) {
+                    const galleryImagePathIndex = product.galleryImagesPaths.findIndex((galleryImagePath) => galleryImagePath === oldGalleryImagePath);
+                    if (galleryImagePathIndex >= 0) {
+                        product.galleryImagesPaths[galleryImagePathIndex] = newGalleryImagePath;
+                        await productModel.updateOne({ _id: productId }, {
+                            galleryImagesPaths: product.galleryImagesPaths
+                        });
+                        return {
+                            msg: "Updating Product Galley Image Process Has Been Successfully !!",
+                            error: false,
+                            data: {},
+                        }
+                    }
+                    return {
+                        msg: "Sorry, This Path Is Not Found !!",
+                        error: true,
+                        data: {},
+                    }
+                }
+                return {
+                    msg: "Sorry, This Product Is Not Exist !!",
+                    error: true,
+                    data: {},
+                }
+            }
+            return {
+                msg: "Sorry, Permission Denied !!",
+                error: true,
+                data: {},
+            }
+        }
+        return {
+            msg: "Sorry, This Admin Is Not Exist !!",
+            error: true,
+            data: {},
+        }
+    }
+    catch(err) {
+        throw Error(err);
+    }
+}
+
+async function updateProductImage(authorizationId, productId, newProductImagePath) {
+    try{
+        const admin = await adminModel.findById(authorizationId);
+        if (admin){
+            if ((new mongoose.Types.ObjectId(authorizationId)).equals(admin._id) && !admin.isBlocked) {
+                const product = await productModel.findById(productId);
+                if (product) {
+                    await productModel.updateOne({ _id: productId }, {
+                        imagePath: newProductImagePath,
+                    });
+                    return {
+                        msg: "Change Product Image Process Has Been Successfully !!",
+                        error: false,
+                        data: {
+                            deletedProductImagePath: product.imagePath,
+                        },
+                    }
+                }
+                return {
+                    msg: "Sorry, This Product Is Not Exist !!",
+                    error: true,
+                    data: product.imagePath,
+                }
+            }
+            return {
+                msg: "Sorry, Permission Denied !!",
+                error: true,
+                data: {},
+            }
+        }
+        return {
+            msg: "Sorry, This Admin Is Not Exist !!",
+            error: true,
+            data: {},
         }
     }
     catch(err) {
